prov/rxm: optimize MSG_EP connect code path and do minor refactoring

+++ prov/rxm/src/rxm_cq.c
@@ -1222,9 +1222,9 @@ static inline int rxm_ep_repost_buf(struct rxm_rx_buf *rx_buf)
 		rx_buf->conn = NULL;
 	rx_buf->hdr.state = RXM_RX;
 
-	ret = fi_recv(rx_buf->msg_ep, &rx_buf->pkt,
-		      rx_buf->ep->eager_limit + sizeof(struct rxm_pkt),
-		      rx_buf->hdr.desc, FI_ADDR_UNSPEC, rx_buf);
+	ret = (int)fi_recv(rx_buf->msg_ep, &rx_buf->pkt,
+			   rx_buf->ep->eager_limit + sizeof(struct rxm_pkt),
+			   rx_buf->hdr.desc, FI_ADDR_UNSPEC, rx_buf);
 	if (ret)
 		FI_WARN(&rxm_prov, FI_LOG_EP_CTRL,
 			"Unable to repost buf: %d\n", ret);
